4th commit; finished and adding encapsulations , final methods and whatnots

+++ src/pointofsale/DiscountStrategy.java
@@ -5,12 +5,13 @@
  */
 package pointofsale;
 
-/**
+/** Description of class
+ * Responsibilities
  *
  * @author Alex
  */
 public interface DiscountStrategy {
 
-    double getTotalAfterDiscount(double price, double qty);
+    double getTotalAfterDiscount(double price, int qty);
     
 }
